5 hours of reconstruction and upgrading
Im starting this upgrade by the first view the user gets when they load the site. I edited out my nav burger and most of my header and reshaped everything in that view. Next step is to finish this and move onto changing the way the site will look when you scroll and the bio so i can have a better chance at getting hired